moved optimal_scales to pygenstability
