+++ src/pygenstability/optimal_scales.py
