+++ src/pygenstability/pygenstability.py
@@ -15,7 +15,7 @@ from tqdm import tqdm
 
 from pygenstability import generalized_louvain
 from pygenstability.constructors import load_constructor
-from pygenstability.contrib.optimal_scales import identify_optimal_scales
+from pygenstability.optimal_scales import identify_optimal_scales
 from pygenstability.io import save_results
 
 L = logging.getLogger(__name__)
